Carousel done

+++ WebSite1/Controllers/ShoppingCartController.cs
@@ -26,7 +26,7 @@ namespace WebSite1.Controllers
             var shoppingCartViewModel = new ShoppingCartViewModel
             {
                 ShoppingCart = _shoppingCart,
-                ShoppingCartTotal = _shoppingCart.GetShoppingCartTotal()
+                ShoppingCartTotal = (decimal)_shoppingCart.GetShoppingCartTotal()
             };
 
             return View(shoppingCartViewModel);
